requirements.txt
openai
langchain
langchain_openai
huggingface_hub
langchain_huggingface
python-dotenv
streamlit
